@@ -1,12 +1,18 @@
 module github.com/trezor/trezord-go
 
-go 1.12
+go 1.24.0
 
 require (
-	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/gorilla/csrf v1.7.0
 	github.com/gorilla/handlers v1.5.1
 	github.com/gorilla/mux v1.8.0
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+)
+
+require (
+	github.com/BurntSushi/toml v0.3.1 // indirect
+	github.com/felixge/httpsnoop v1.0.1 // indirect
+	github.com/gorilla/securecookie v1.1.1 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
